Bug fix and usability improvement related to fancybox.
The new fancybox-based dialogs have had arbitrary fixed sizes and were
not resizable by the user which leads to poor usability.  This has been
addressed in part by upgrading to fancybox 1.3.4 which allows for specifing
the dialog dimensions as percentages of those of the containing frame.

Also a bug was fixed where restoreSession() was not called when
the dialog was opened, causing possible conflicts between two login
sessions on the same workstation.

These changes currently apply only to the New Encounter form.
Similar corrections will be needed in other places.