Do not include the empty demographic fields in the reports when they are empty.
Problem Summary:
Patient report generation was printing all the demographics fields even if they were empty.

Description:
There are like around 50 fields in demographics and the patient report prints all the 50 fields with empty labels even if we fill just 3 or 4.

Fix:
I am filtering the empty fields not to be included for printing in the reports.

--Contributed by Athiyaman Muthukumarasamy.