Fix to using most recent development translation set on development demo.
Github can not serve text files are 10MB, so needed a work around. The
development demo now serves this file; this is not ideal since it goes down
for about 20 minutes a day while resetting, but it at least works for now.
Will also direct the wiki link for most recent translation development set
to this file.