Update sql_upgrade.php
This code meets:
1.When a sql "version" table does not exist, then continue the current behavior (ask the user for the version).
2.When a sql "version" table does exist, then calculate the version and run the upgrade script without any user input. There will be two situations in this case:
scenario 1: More common scenario where it is a production version(no "-dev" in the version table in the v_tag column). In this case, can just use the version.
scenario 2: Less common scenario where it is a development version (with "-dev" in the v_tag column). In this case, will need calculate the actual version by integrating the version in the version table with the upgrade files in sql/ (you will need to go to one version below what is in the version table; note you can do this quickly by using the $version array which is already being created in the current sql_upgrade.php script at line 37).