Toggle Construction zones
As the CC,PHI,PMFSHx are being developed I chose to hide these
preliminary fields for now.  As they are developing you can toggle
their visibility but for now they are a distraction to the main
clinical portion of the form.  Perhaps other forms can be incorporated
within openEMR rather than reinvent the wheel?