Menu Behavior Tweaks
The experimental menu look and feel is nearly complete.
You can only see it at present by clicking “+” sign top right of an
eye_mag form.
The behavior mimics the menu on my Mac so I assume that is OK with
others.  Will tweak per suggestions.
Need to seed the menu with actual links to OpenEMR.
HPI quick picks paved out.  Logic to build HPI narrative from
contextual fields yet to be done.  Phrase building is not an exact
science so the logic to do this is still fuzzy to me.  Simple building
blocks are easy but locale and language dependent.  Will work on it
when I get clarity. Time to hit the moguls with the kids.
IMP/PLAN builder is still empty…