Fix for the recurring appointment bug in Calendar
This fixes a recurring appointment bug in the Calendar, which caused
both missing appointments and appointments with wrong date. It effected
these recurring appointments...
every (1st, 2nd, ...) (Mon, Tue ...) of the month
set on start date 29, 30, or 31.
Example - setting a recurring appointment for every last Friday on the
31/12/2014 would result in the appointment for Feb going missing.
If the appointment was set for every 2nd month, then all dates were
wrong after December.