Created the About page, take 2.
The About link was added to top right (in place of the Manual link).
The 'Online Support' and 'Manual' links were migrated to the About popup.
The version information was migrated from the login screen to the About popup.
A option to show a support phone number was added to Globals that will show in the About popup.