Added code to group appts with same date into sets for display in the Patient Summary, take 2.
Added code to group appts with same date into sets for display in the Patient Summary

When the number of appts scheduled is greater than the number of appts configured to appear in the Appointments widget the last set turns green if all appts for the corresponding date are included for display and pink if all appts are not present.

Also two fixes...

Fix 1.

If the Calendar (including the Appointments widget) is deactivated the Clinical Reminders widget is disabled also (it was within the scope of the if statement that controls whether the Appointments code executes).

Fix 2.

When an appt is opened for editing in the Past Appointment widget fields contain either no data or wrong data (parameter list mismatch between call to oldEvt from Past Appointment code and the function itself caused when I changed oldEvt to support recurring appts in Appoinments).