Fix - missing and incorrect dates for rt2 recurring appointments set on the 29th, 30th and 31st (#400)
For rt2 recurring appointments set on the 29th, 30th and 31st where there is a month associated with a date in the corresponding recurring series that has a total number of days less than the day of month the appointment was set on (29 30 or 31) (first month if more than one) the date for this month will be missing when freq = 1. For freq > 1 all subsequent dates will be missing also and a series of wrong dates will be included each one month shifted from the missing correct dates.

Dates for an rt2 appointment set on 2016-12-31...

freq = 1
2016-12-31
2017-01-28
2017-03-25
2017-04-29
2017-05-27
2017-06-24

freq = 2
2016-12-31
2017-03-25
2017-05-27
2017-07-29
2017-09-30
2017-11-25

Only new (post fix) appointment experience the fix, old appointments remain unchanged.