Logout fix. (#420)
This is to prevent following logout bug:
1. Login to OpenEMR.
2. Go to Patient Flow Board and toggle on Open Patient in New Window.
3. A new window(OpenEMR instance) now opens up.
4. Click to logout of the new OpenEMR instance.

Because of code in library/auth.inc, the following bugs happen:
1. The window of the OpenEMR instance that you clicked to logout
   will go away.
2. The Original window (from step 2 above) will then logout.

This is because the 'opener' javascript variable is linked to the
original instance. The fix is to always set the 'opener' variable
to null in the main script for frames (interface/main/main_screen.php)
and tabs (interface/main/tabs/main.php).