Highway to PSR2
PSR2 rules fully expanded into phpcs.xml, with all errors commented out and marked as fixable or as targets for warnings.

Removing `?>` from files with only php

"There MUST NOT be trailing whitespace at the end of non-blank lines."

When present, there MUST be one blank line after the namespace declaration.

When present, all use declarations MUST go after the namespace declaration.
There MUST be one use keyword per declaration.
There MUST be one blank line after the use block.

The extends and implements keywords MUST be declared on the same line as the class name.
The opening brace for the class go MUST go on its own line; the closing brace for the class MUST go on the next line after the body.
Lists of implements MAY be split across multiple lines, where each subsequent line is indented once. When doing so, the first item in the list MUST be on the next line, and there MUST be only one interface per line.

Visibility MUST be declared on all methods.

method visibility precedes static keyword and follows abstract keyword

(Mostly removing extra lines before closing braces in methods.)
Method names MUST NOT be declared with a space after the method name.
The opening brace MUST go on its own line, and the closing brace MUST go on the next line following the body.
There MUST NOT be a space after the opening parenthesis, and there MUST NOT be a space before the closing parenthesis.

In the argument list, there MUST NOT be a space before each comma, and there MUST be one space after each comma.

When making a method or function call, there MUST NOT be a space between the method or function name and the opening parenthesis, there MUST NOT be a space after the opening parenthesis, and there MUST NOT be a space before the closing parenthesis.
In the argument list, there MUST NOT be a space before each comma, and there MUST be one space after each comma. (**MOSTLY THIS**)
Argument lists MAY be split across multiple lines, where each subsequent line is indented once. When doing so, the first item in the list MUST be on the next line, and there MUST be only one argument per line.

Argument lists MAY be split across multiple lines, where each subsequent line is indented once. When doing so, the first item in the list MUST be on the next line, and there MUST be only one argument per line.

There MUST be one space after the control structure keyword
 There MUST NOT be a space after the opening parenthesis
 There MUST NOT be a space before the closing parenthesis
 There MUST be one space between the closing parenthesis and the opening brace
 The structure body MUST be indented once
 The closing brace MUST be on the next line after the body

The body of each structure MUST be enclosed by braces. **No inline control structures**

The case statement MUST be indented once from switch, and the break keyword (or other terminating keyword) MUST be indented at the same level as the case body.
There MUST be a comment such as // no break when fall-through is intentional in a non-empty case body.

Added scope indentation rules. About 50 files needed manual fixes, therefore **this commit needs manual inspection and testing**.
Additionally adds foreach rules.

Running CI with updated phpcs. Had to disable ScopeClosingBrace rule