Boards: Today button works as expected
Default datetimepicker “Today” button onclick just highlights current date.   That’s it.  Now it actually selects Today’s date and fills in the input field and closes the datepicker.  This is the expected behavior (but it is not).  This code would need to be added to each file including the datetimepicker desiring this functionality.

Signed-off-by: ophthal <magauran@ophthal.org>