Removing unnecessary frameset and related bugs. (#1110)
* Removing unnecessary frameset and related bugs.

When an encounter is entered the code has been creating a frameset for it with
a single frame named "Forms". This seems to be a holdover from when the frameset
had a second frame in it.

Currently this frameset seems unnecessary since it is contained in another frame
or iframe, depending on the layout. In addition there is at least one bug where
canceling forms causes the frameset to be nested to arbitrary levels.

This commit removes the "Forms" frameset.