User Experience tweak
When deleting a document, a modal pop-up asks the user to confirm the action, warning that the action is logged.  On choosing yes, the Modal changes to show mysql data and then a js alert pops-up to confirm the file was deleted.

This change eliminates the last two steps - once the user selects “Yes” to delete the document, the modal simply closes.  No need for a deletion confirmation via ls alert.

Signed-off-by: ophthal <magauran@ophthal.org>