Documents list
This will make the left side of the documents window resizable, allowing the user to “drag” the divider to increase or decrease the size of the panes.  This is useful when a document has a very long name and the user would like more space/real estate to view a document inline.

Signed-off-by: ophthal <magauran@ophthal.org>