Installer Bug fix
The installer script fails at OpenEMR Setup Step 3 on a new installation of openEMR
if the user already exists in mysql. This is caused by the public function create_database_user()
that was created to allow the script to execute in mysql8.

In order to account for an existing user a new function has been added public function check_database_user()
that will check to see if user exists and will add user only if it does not exist

openemr/setup.php
openemr/library/classes/Installer.class.php