Modified form vitals with suggestions in PR comments
Added background to date-time input for current vital to highlight that
this is the current date-time regardless of encounter date. Not realising this
can lead to delayed entry of vitals reflecting the date-time it was entered,
care needs to be taken to change date-time as appopriate.

Modified interface/patient_file/encounter/trend_form.php to get rid
of a bug that caused the graph to not display if it was selected after
selecting an item that had no data. Included an alert that is informative.