NPI Validation
NPI validation uses the Luhn algorithm (used to verify credit cards).  The NPI number must be prepended with 80840.  This is a modification to perform NPI validation in javascript.

In reality I just extended  @author ShirtlessKirk validation code which is the cleanest implementation of the Luhn algorithm I could find, and it was already in the codebase.

Signed-off-by: ophthal <magauran@ophthal.org>