Most MOE
This is a working prototype for mobile devices.
It allows users to search through Documents for a patient, add new documents, add new photos and add new videos directly into theirEHR without saving the images on the local device.
It display calendars for multiple providers.
There is a mobile version of Flow Board.
The SMS Bot is fully integrated (MedEx only).

Signed-off-by: ophthal <magauran@ophthal.org>