Added layout edit options J, K, M and m.
J is to start this field on a new row, regardless of whether the current row is full.
K is to insert a blank row before this field, and also forces a new row.
M and m work with radio button fields. A field with M is a "master" field which, when changed,
will cause the following m "member" fields to receive the same change. You can have multiple
such groups with each M field starting a new group.