Removing InnerHtml and updating text content
Removing Inner Html, and Updating with Text Content.  I attempted to send a post request on my local machine to find this specific end point, and see the return data. So long as return data is not any Html to be parsed this should update it. If there is html, we can change the post response, or add the elements to the DOM, prior to the text content.