Fixes #2324 Rest API Fatal Error. (#2325)
* Fixes #2324 Rest API Fatal Error.

The Rest API was fatally failing if you sent an invalid site id.  The code was using the escape function of text() but
that file is not loaded until later on in the codebase.  I could have moved the load order of the files so that
composer (which in turn loads htmlspecialchars.inc.php) would load earlier, but this appeared to be the safer approach
to minimize any system changes by just calling htmlspecialchars(..) directly.

Also documented what the scope parameter is supposed to be in the REST API as it was unclear with the example 'site id'
for a newcomer to the system.

* Fix 2324 Change error message to restrict details

To avoid giving external users information about what went wrong we just say invalid url.
We log the details of it to the error log so developers know what is going on.

* Fixes #2324 fix code linting errors.

Used the wrong editor to make the last commit and had incorrect spacing for the project.
Fixed that.