Pull modules from database. Fix error logs.
Fixed some extra error logs showing up when the class objects are not found.
Useful for debugging but throws a lot of extra log entries.

Reconfigured the application config file for the zend modules to pull the zend modules
from the database.  It looks like I missed something in the Uninstalled view file where
it does a directory discovery to find modules and then install them.  So if you add folders
with modules in the directories they will get inserted into the database when you 'register'
them.

Since zend is loaded after the modules are loaded we can't use the zend adapters to hit the
database.  So we do a sqlStatement just before to grab the loaded module definitions.