Zend Module Event Dispatcher w/ Menu Example
Event Dispatcher integration with Tab Menu system

Made it so we integrate the event dispatcher code in the OpenEMR system with the ZF3 module code.

Moved the module menu code outside of the core OpenEMR codebase into the module system which is where
the code should belong as it creates and defines the module database code.

Added the Update and Restrict menu events (defined in Menu/MenuEvent) and tied them into the tabs interface
as an example of how to use the EventDispatcher to hook into the main codebase to extend it.

All of this moves the code to be more S.O.L.I.D (https://en.wikipedia.org/wiki/SOLID) by modifying the module menu system.
The menu files in Core are closed for modification but Open for extension as well as better moving to Single Responsibility (SRP).

Change side nav to use tabs module code

The side navigation uses a different codebase that uses inline code instead of in memory objects for the initial menu structure.
Refactored the module code to leverage the existing Zend Module listener code that the tabs menu system uses.

This de-duplicates the module hooks code and centralizes it into the zend modules codebase.

This also illustrates the effectiveness of the event dispatcher as multiple places in the code can fire off the event and the same
listeners can respond appropriately which de-couples the systems.