Exchange float:left for oe-pull-toward-unimportant
oe-pull-toward-unimportant (and as well as away) has been added
to both rtl_style_pdf.css file and the help-files-common.scss file.

It has the same functionality as the one without the 'unimportant',
but simply does not have the !unimportant tag with it. This is so
that the further dropdown menus are able to have the "float:none",
allowing them to function as normal.

The class has been added to the menu_template.php file and the
user_data_template.php file.