Add horizontal scroll to PostCalendar Day template (#2706)
If you have many providers, the calendar view only shows the
first however many fit on the page (depending on your screen
size). If you have a large number of providers, the only way
to see their schedules is to select individuals in the
"pc_username" box.

When blocking out time for vacations, it is useful for us to
be able to get a quick overview of what is going on with all
the providers.

This just adds "overflow-x: auto;" to the "bigCal" div style
on the Day template, allowing us to view the provider schedules
that render past the right-hand limit of the div.