Weno new standard build update
The foundation has been laid for the new weno standard for eRx

Adding validation back to the process and ...

Forgot to remove the patient chart pharmacy entry

can't get rid of this file

Added the spinner and cleaned up the response

Waiting for radification from Weno

minor adjustments

Updates

Revisions done per requests

PHPCBF finally got it installed and working in PHP. Code should be beautiful now

This time I set the standard to PSR2.

Third times a WTF moment in time

Added a few notes to these two files and a filter input to the GET.

accidents happen

New round of views and completed this set

PSR2 corrections

A few corrections needed to be made.

here are all the changes that are in the patch that are not in the repo

Revert "here are all the changes that are in the patch that are not in the repo"

This reverts commit 0f1f242151d256663a22f1a1c2f65762a3ddfccc.

Revert "here are all the changes that are in the patch that are not in the repo"

This reverts commit 0f1f242151d256663a22f1a1c2f65762a3ddfccc.

Signed-off-by: Sherwin Gaddis <sherwingaddis@gmail.com>