Move deletion of issues to the table
Preiously, to delete an issue from the table, one opened the Add/Edit Issue dialog
which contained a Delete button at the bottom. This workflow is non-standard and a
but non-intuitive.

Now the issues table contains a Delete button with a checkbox column to select
the issue row(s) to delete. The column header's checkbox toggles
selection/de-selection of all the rows. The Delete button in the
Add/Edit Issue dialog has been removed.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>