SMART launch/patient scope, CORS enabled, logging
Bunch of work to enable the standalone SMART app launch with the
launch/patient scope for SMART.

This replaces the current IdTokenResponse class used in the oauth2
authentication process with an IdTokenSMARTResponse which will be
able to add on context dependent data that is sent back to the client
parallel to the accessToken and idToken. Right now it just grabs
the very first patient available in the system.  It needs to be
wired up so that a provider gets a patient selector or when a
patient logs in it sends that patient information as part of
the external app launch

This also enables CORS support for the browser based oauth2 requests
and sets the OPTIONS requests to bypass any token validation / support.

Implemented a PSR3 logger which wraps around the Monolog logger package.
Wrote a Decorator/Adapter class around the Monolog logger so the system
isn't tightly coupled to Monolog and we can replace it if its ever
needed.

The logging needs to be tied to somekind of global config so we can
decide the log level as needed.  These logs are currently NOT intended
for system level auditing, but writes directly to the error log.
Anything higher than the currently set system log level is treated like
a noop.

The logging helped immensely in tracking down OAUTH2 problems.