SMART EHR launch context support
Implemented the 'launch' SMART Core Context for in EHR launching with a patient selected.

Added a display of the registered SMART apps as part of the patient demographics.  Right now
this is a placeholder of where we can put it to generate ideas of what we can do to display
the SMART apps.  It filters out from the registered client applications ONLY those that
have registered as a SMART application by providing the 'launch' scope as part of their
initial registration.

It assumes that the registered app has a 'launch.html' file which appears to be pretty widespread
standard.  However, we may need to add a redirect_launch_uri or something.  The spec is vague and
needs to be dug into a bit to see if there is a standard, or if there's an industry best practice
here.