SMART Core Capability context_style,context_banner
Implemented the two smart core capabilities for the banner and context
style.

Created a basic developer app registration.  Right now it registers
everything as a smart app, will need to provide developers the option of
specifying what permissions they need.  Currently it directly enables
the developer apps, but an admin panel will need to be setup to
enable/disable the dynamic registration for the developer apps.
After a registration is successful the client id is returned to the
registrant.

Implemented pulling up the SMART app in a dialog from the patient
demographics page.

You can test this out by running a modified version of the Growth Charts
open source app hosted at https://adunsulag.github.io/growth-chart-app/

Register a client with the redirect_uri to be:
https://adunsulag.github.io/growth-chart-app/

and the logout uri to be https://adunsulag.github.io/growth-chart-app/logout

Open a patient and at the bottom of the demographics you will now have a
Launch App button.  Launch the app and enter in the Client ID from your
registration process.

If you need to clear the client_id for testing you can do so at
https://adunsulag.github.io/growth-chart-app/deregister.html