Confidential/SMART app reg is_enabled flag
Made it so that an app that displays to the user in the patient
demographics must have the is_enabled database flag set to true.  This
prepares the way for us to build an admin panel that will turn on and
off the SMART apps.  Developers can still build against the OpenEMR
install in preparation for MU3 but they have to manually switch the flag
in the database for their SMART app to show up in patient demographics.

Any confidential app or app requesting the 'launch' scope will require a
manual approval from the OpenEMR administrator.  So apps can still
dynamically register but it's a two step process.