Basic OAUTH2 Admin Client Management
Added a very basic client admin system for creating and approving
oauth2 apps.  It uses the app-admin.php as the main entry point and then
launches everything into a Frontend Controller which is
ClientAdminController.

Since there are only 3 actions and I'm not sure
if we are still wanting to use twig/smarty the class also functions as
the Presentation class in addition to being a dispatch controller. So
the html is mixed with the php.  We should be able to easily separate
this out if we need to later.

Still not sure on how the javascript system is setup in OpenEMR so I'm
just passing message responses back following an HTTP Request/Location redirect
pattern for saving data.