Smart fhir admin enable / disable management (#4142)
* SMART admin page.

* Confidential/SMART app reg is_enabled flag

Made it so that an app that displays to the user in the patient
demographics must have the is_enabled database flag set to true.  This
prepares the way for us to build an admin panel that will turn on and
off the SMART apps.  Developers can still build against the OpenEMR
install in preparation for MU3 but they have to manually switch the flag
in the database for their SMART app to show up in patient demographics.

Any confidential app or app requesting the 'launch' scope will require a
manual approval from the OpenEMR administrator.  So apps can still
dynamically register but it's a two step process.

* Revert debug flag on logger.

I wonder if we should add to our globals a dropdown that the systemlogger will pull
from or something.  This is way too easy to casually forget what logging
level we are set to.

* Basic OAUTH2 Admin Client Management

Added a very basic client admin system for creating and approving
oauth2 apps.  It uses the app-admin.php as the main entry point and then
launches everything into a Frontend Controller which is
ClientAdminController.

Since there are only 3 actions and I'm not sure
if we are still wanting to use twig/smarty the class also functions as
the Presentation class in addition to being a dispatch controller. So
the html is mixed with the php.  We should be able to easily separate
this out if we need to later.

Still not sure on how the javascript system is setup in OpenEMR so I'm
just passing message responses back following an HTTP Request/Location redirect
pattern for saving data.

* Enable / Disable OAUTH2 clients.

You can now enable / disable an oauth2 client registration. Added instructions
and admin page cleanup.

* Fix styling issues.

* Oops didn't get this checked in.

* Fix session difference w/ top window.

Did a bit more testing, we were losing session context somehow and I
wasn't sure why.  Figured out looking at other scripts that we need to
do top.restoreSession on every href click.  So we have that here.

* Attempting to fix style issues.

So local environment isn't reporting any style issues but the live
system is. Not sure what is going on.  Tried fixing it off what the
github CI says, we'll see if this works.

* Remove custom menu piece with invalid script

The custom menu wasn't doing anything anyways so we remove the menu item
from it.

Co-authored-by: Stephen Nielson <snielson@discoverandchange.com>