OAUTH2 Standalone Provider Patient Selector
Implemented a patient selector for oauth2 standalone launch.

Also made it so we could have some primitive unit tests on the Gacl
class with resetting the ACL cache.

Wrote a smart auth class for any context / ui needs that are required
as part of the oauth2 flow.

Eventually if ONC requires the context-launch-encounter piece we can
easily add to this class to facilitate additional selector flows.

Implemented ACL checks and a search class to be used as part of the
oauth2 flow for patients.