Centralized scope checks for rest api. Perm update
Updated the scope permissions so that only the ones we currently support
for patient/<resource>.* is supported.

Centralized the scope checks in the HttpRestRouteHandler to remove
redundant code from the routes and to make sure all possible routes are
checked against the access code.  This also prevents developers who
extend the api from accidently forgetting to check against the
AccessToken.

For now I've only enabled it on the FHIR api, but if it tests well we
can open it up to the rest of the APIs.