Inferno Limited Scope Authorization
Implemented a logged in user being able to restrict what scopes they are
giving access to for the requesting application.  This allows OpenEMR
users to prevent an application from giving offline_access (credentials
past the 1 hour access token), and other resources to an application
that they don't want to provide.