Bulk FHIR Export authentication,api requests
Initial Implementation of the Export Controller to handle Bulk FHIR Export.

Put in the initial $export routes for the Group/Patient/System exports of
data for the BULK FHIR implementation.

Refactored the Capability statement to now support FHIR operations on each
FHIR resource.  Moved the Capability statement to use the FHIR class
resources for the REST response.  This fixed a number of issues in the FHIR
validator for inferno.

Pulled out the oauth2_trusted_user api calls into its own controller so we
can grant a trusted user for Client Credentials grant.

Refactored dispatch.php in order to allow a trusted client access token that
is not tied to a user to still have api access for system calls.

Refactored the hard dependancy on RestConfig in the ScopeRepository so I could
write a unit test against ScopeRepository to fix some scope bugs that were
occurring.