SMART FHIR Group Compartment Export
Got the SMART FHIR Group Compartment Implementation Guide works.

Added the SystemUser to the UserService and got it so the Client
Credentials grant uses the SystemUser and is checked properly in the api
dispatch file.

Wrote a wrapper class around our embedded Psr17Factory so we can replace
it as needed.

Created a StatusCode class that we can start putting HTTP status in
which makes unit testing and readability much better.

Wrote an ExportJob class to match with the database export_job and that
the FhirExportRestController creates and manages the life cycle of.

Wrote the controller class to properly handle the bulk $export and the
$export status checks.

Right now the file creation of the exports and the deletion of the
exports is mocked out and need to be implemented in a subsequent
iteration.

Minor style guide fixes included as well.