USCIG core fixes to Patient FHIR resource
Added required address period field.  Partially implemented ethnicity and race
extensions. Added required communication field.

The FHIR specification is inconsistent in the SystemURIs for the
language communication, ethnicity and race.  Different documents state
different URIs for this and I can't seem to find what the correct one
is.  For now its just throwing warnings in inferno and not failing the
test.  We'll need to figure out the correct one, though it seems like
its a moving target as some of the checker's changed as of mid January.