Add foreign db key to Documents. Fix document bugs.
Added a foreign reference id & table name to the Documents class so we
can tie a Document to another table record.  Apparently the 'foreign_id'
only is used for Patients per @bradymiller.

I also fixed a bug where the deleted flag was not actually picked up by
the document ORM since it was missing the getters and setters.

Finally fixed a bug in the FhirDocumentRestController that was allowing
access to deleted documents that had not yet expired.